Adding flexibility over the path for the chronyd UDS socket
This commit ensures that one is able to make calls over UDS over a
customizable path. This does not break the existing implementation and
keeps it backward compatible, while snuring that it gives us more
control on directing towards a customized file path.

There's an additional dependency brought on a tempfile create and moving
the client socket into a temp directory, but this isn't a breaking
change.